2021-09-17 (#5)
* Added support for Star Wars Saga

* Updated version number

* Update MythWeaversStatblock.user.js

* Update MythWeaversStatblock.user.js

* Conditional expressions, nesting

Added conditional expression, support for nesting.

* Added conditional expressions

* This was a test, removing as it's no longer needed

* Whitespaces in conditional expressions

* Added math

* Added math

* Named groups in regex

Improved to use named groups in regular expressions rather than group numbers. Easier to read, less likely to break if additions are made to the regular expressions.

* Subsections, aliases

Added identifier subsections.
Added aliases.

* Default identifier, additional math

If no identifier is given (as in: ::[section]::), we assume that we're meant to look into the private notes.
Added the FLOOR and CEILING extras to math expressions.

* Started writing walkthrough, not finished

* Added Mathematics, started Section

* Expanded the use of Sections

* Rewrote as a tutorial

* Empty file - was needed to create the folder

* Beta (#3) (#4)

* Added support for Star Wars Saga

* Updated version number

* Update MythWeaversStatblock.user.js

* Update MythWeaversStatblock.user.js

* Conditional expressions, nesting

Added conditional expression, support for nesting.

* Added conditional expressions

* This was a test, removing as it's no longer needed

* Whitespaces in conditional expressions

* Added math

* Added math

* Named groups in regex

Improved to use named groups in regular expressions rather than group numbers. Easier to read, less likely to break if additions are made to the regular expressions.

* Subsections, aliases

Added identifier subsections.
Added aliases.

* Default identifier, additional math

If no identifier is given (as in: ::[section]::), we assume that we're meant to look into the private notes.
Added the FLOOR and CEILING extras to math expressions.

* Started writing walkthrough, not finished

* Added Mathematics, started Section

* Expanded the use of Sections

* Rewrote as a tutorial

* Empty file - was needed to create the folder

* Stripped math expression

Remove everything from math expression that isn't a number, a dot (decimal separator), a parenthesis, or an operation (+-*/)

* Removed mathjs library

Since we sanitize expressions for the MATH() tag, we no longer need the mathjs library. Also added optional + sign for those MATH tags.

* Added more in maths

Added more options / explanation for the math section.

* Added contact info